Split test_configure tests into 2 test modules
